requirements.txt
python-decouple
motor[srv]==3.7.1
dnspython
lxml_html_clean
aiohttp
aiofiles
apscheduler==3.10.1
feedparser
telethon==1.40.0
pyrofork==2.3.67
tgcrypto
AnilistPython==0.1.3
anitopy
git+https://github.com/kaif-00z/html-telegraph-poster
pytz
Flask
